requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-dotenv
jinja2
requests
python-multipart